Update README.md
A hotel management system in which customers can see the available number of hotel rooms and the number of food items available. The food items include burgers, Pasta, Noodles, Chicken-rolls, and Shakes. The hotel manager can therefore see the status of total sale collections for the day.

The manager can feed the number of available items in the hotel and then can accordingly do the required. The customers would have to select the key number for the product they want to buy. After the selection they will get the option for choosing the quantity of that product, also the price of each item is also displayed so that they don't have to face any kind of inconvenience as per their budget. There is a little limitation for sale, if a person wants to make a purchase for more than one item, then they can only select the preceding options of items that are left. items are put in an order, if they buy a product that lies mid of the list, they can't buy the items that are above of it, but they can do the same if it lies below it.